Make a more logical off setting.
The reason that we changed off to false is because we listen for the
true input.

If `full=true` or full is not defined then we use pywal
Otherwise pywal will be disabled.
In our case full=off) you would think to use `full=on`  which would
still disable pywall

Thus we changed off -> false